+++ telepot/helper.py
@@ -161,6 +161,8 @@ class Administrator(object):
     - :meth:`.Bot.getChatAdministrators`
     - :meth:`.Bot.getChatMembersCount`
     - :meth:`.Bot.getChatMember`
+    - :meth:`.Bot.setChatStickerSet`
+    - :meth:`.Bot.deleteChatStickerSet`
     """
 
     def __init__(self, bot, chat_id):
@@ -179,7 +181,9 @@ class Administrator(object):
                        'getChat',
                        'getChatAdministrators',
                        'getChatMembersCount',
-                       'getChatMember',]:
+                       'getChatMember',
+                       'setChatStickerSet',
+                       'deleteChatStickerSet']:
             setattr(self, method, partial(getattr(bot, method), chat_id))
 
 
@@ -197,6 +201,8 @@ class Editor(object):
     - :meth:`.Bot.editMessageCaption`
     - :meth:`.Bot.editMessageReplyMarkup`
     - :meth:`.Bot.deleteMessage`
+    - :meth:`.Bot.editMessageLiveLocation`
+    - :meth:`.Bot.stopMessageLiveLocation`
 
     A message's identifier can be easily extracted with :func:`telepot.message_identifier`.
     """
@@ -214,7 +220,9 @@ class Editor(object):
         for method in ['editMessageText',
                        'editMessageCaption',
                        'editMessageReplyMarkup',
-                       'deleteMessage']:
+                       'deleteMessage',
+                       'editMessageLiveLocation',
+                       'stopMessageLiveLocation']:
             setattr(self, method, partial(getattr(bot, method), msg_identifier))