+++ test/test3_updates.py
@@ -87,6 +87,8 @@ def see_every_content_types(msg):
 TOKEN = sys.argv[1]
 USER_ID = int(sys.argv[2])
 
+# telepot.api.set_proxy('http://192.168.0.103:3128')
+
 bot = telepot.Bot(TOKEN)
 
 expected_content_type = content_type_iterator.__next__()